fix(admin-id/controller): prunning
throw error if the user doesnt exist

+++ src/features/user/admins/id/admin-id.controller.ts
@@ -10,33 +10,41 @@ export class Controller {
   modifyUserInfo: RequestHandler = async (req, res, next) => {
     console.log("received from Client");
     const { id } = req.params;
-    const user = req.user
-    const userId = user?.id
-    if (!userId) throw new HttpError(401,"unauthorized") 
-    
-    await service.modifyUserInfo(userId)
-    
+    const user = req.user;
+    const userId = user?.id;
+    if (!user) throw new HttpError(400, "해당 유저가 없습니다");
+    if (user.role !== "SUPER_ADMIN")
+      throw new HttpError(403, "권한이 없습니다");
+    if (!userId) throw new HttpError(401, "unauthorized");
+
+    await service.modifyUserInfo(userId);
+
     return res.status(204).json();
   };
   modifyJoinStatus: RequestHandler = async (req, res, next) => {
     const { id } = req.params;
     const { joinStatus } = req.body;
 
-    const user = req.user
-    const userId = user?.id
+    const user = req.user;
+    const userId = user?.id;
 
-    if (!userId) throw new HttpError(401,"unauthorized") 
-    await service.modifyStatus( userId, joinStatus)
+    if (!user) throw new HttpError(400, "해당 유저가 없습니다");
+    if (user.role !== "SUPER_ADMIN")
+      throw new HttpError(403, "권한이 없습니다");
+    if (!userId) throw new HttpError(401, "unauthorized");
+    await service.modifyStatus(userId, joinStatus);
     return res.status(204).json();
   };
   deleteAdmin: RequestHandler = async (req, res, next) => {
     const { id } = req.params;
-     const user = req.user
-    const userId = user?.id
+    const user = req.user;
+    const userId = user?.id;
 
-    if (!userId) throw new HttpError(401,"unauthorized") 
-    await service.deleteAdmin( userId )
+    if (!user) throw new HttpError(400, "해당 유저가 없습니다");
+    if (user.role !== "SUPER_ADMIN")
+      throw new HttpError(403, "권한이 없습니다");
+    if (!userId) throw new HttpError(401, "unauthorized");
+    await service.deleteAdmin(userId);
     return res.status(204).json();
-
   };
 }
